tests/wpt/web-platform-tests/css/css-text/parsing/text-align-all-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-align-all with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align-all">
<meta name="assert" content="text-align-all supports the full grammar 'start | end | left | right | center | justify | match-parent'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-align-all", "start");
test_valid_value("text-align-all", "end");
test_valid_value("text-align-all", "left");
test_valid_value("text-align-all", "right");
test_valid_value("text-align-all", "center");
test_valid_value("text-align-all", "justify");
test_valid_value("text-align-all", "match-parent");
</script>
</body>
</html>
