third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-align-all-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-align-all'] = "start" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "end" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "left" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "right" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "center" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "justify" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-align-all'] = "match-parent" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
